requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
Pillow==9.2.0
sqlparse==0.4.2
gunicorn>=20.0,<21.0
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0
